% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{add_geometry}
\alias{add_geometry}
\title{Add a geometry column to a cell seg table}
\usage{
add_geometry(csd)
}
\arguments{
\item{csd}{A cell seg data table}
}
\value{
The table with a \code{geometry} column of \code{sf::st_point} objects
}
\description{
Add a geometry column to a cell seg table
}
