% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{compute_all_nearest_distance}
\alias{compute_all_nearest_distance}
\title{Nearest neighbors from a file}
\usage{
compute_all_nearest_distance(cell_table_path = NULL, out_path = NULL)
}
\arguments{
\item{cell_table_path}{Path to an inForm cell seg data file, or NULL
to prompt for the path.}

\item{out_path}{Path to the output file, or NULL to create a path from the
input file path.}
}
\description{
Compute nearest distance to each phenotype for each cell in a
(possibly merged) inForm cell seg table. Write the result to a new file.
}
\seealso{
Other distance functions: \code{\link{average_count_within}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{subset_distance_matrix}}
}
