% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_touching_cells_fast.R
\name{count_touching_cells_fast}
\alias{count_touching_cells_fast}
\title{Count touching cells for a single pair of phenotypes}
\usage{
count_touching_cells_fast(csd, field_name, export_path, phenos, color1,
  color2, discard_dups = FALSE)
}
\arguments{
\item{csd}{Cell seg data, may include multiple fields}

\item{field_name}{Sample Name or Annotation ID to process}

\item{export_path}{Path to a directory containing composite, component,
and segmentation image files from inForm}

\item{phenos}{Named list of phenotype definitions. Must have length 2.}

\item{color1, color2}{Colors to draw the phenotype dots}

\item{discard_dups}{If TRUE, cells matching both phenotypes will be
discarded from the output. If FALSE, overlapping phenotypes is an error.}
}
\value{
Returns a `list` containing two items:
\describe{
  \item{`image`}{An `EBImage::Image` of touching cells.}
  \item{`data`}{A \code{\link[tibble]{tibble}} containing
  data about the touching cells.}
}
}
\description{
Fast implementation, requires new-style (label image) membrane map.
}
