% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_touching_cells_single.R
\name{count_touching_cells_single}
\alias{count_touching_cells_single}
\title{Count touching cells for a single pair of phenotypes}
\usage{
count_touching_cells_single(csd, field_name, export_path, phenos, color1,
  color2)
}
\arguments{
\item{csd}{Cell seg data}

\item{field_name}{Sample Name or Annotation ID to process}

\item{export_path}{Path to a directory containing composite and component
image files from inForm}

\item{phenos}{Named list of phenotype definitions. Must have length 2.}

\item{color1, color2}{Colors to draw the phenotype dots}
}
\value{
Returns a `list` containing two items:
\describe{
  \item{`image`}{An `EBImage::Image` of touching cells.}
  \item{`data`}{A \code{\link[tibble]{tibble}} containing
  data about the touching cells.}
}
}
\description{
Count touching cells for a single pair of phenotypes
}
