% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within}
\alias{count_within}
\title{Count cells within a radius}
\usage{
count_within(csd, from, to, radius, category = NULL, dst = NULL)
}
\arguments{
\item{csd}{A data frame with \code{Cell X Position},
\code{Cell Y Position} and \code{Phenotype} columns,
such as the result of calling \code{\link{read_cell_seg_data}}.}

\item{from, to}{Selection criteria for the
rows and columns. Accepts all formats accepted by \code{\link{select_rows}}.}

\item{radius}{The radius or radii to search within.}

\item{category}{Optional tissue category to restrict both \code{from} and
\code{to}.}

\item{dst}{Optional distance matrix corresponding to \code{csd},
produced by calling \code{\link{distance_matrix}}.}
}
\value{
A \code{\link{data_frame}} with four columns and one row for each
  value in \code{radius}:
  \describe{
   \item{\code{radius}}{The value of \code{radius} for this row.}
   \item{\code{from_count}}{The number of \code{from} cells found in
    \code{csd}.}
   \item{\code{to_count}}{The number of \code{to} cells found in \code{csd}.}
   \item{\code{from_with}}{The number of \code{from} cells with a
   \code{to} cell within \code{radius}.}
   \item{\code{within_mean}}{The average number of \code{to} cells found
   within \code{radius} microns of each \code{from} cell.}
 }
}
\description{
Count the number of \code{from} cells having a \code{to} cell within
\code{radius} microns in tissue category \code{category}.
Compute the average number of \code{to} cells
within \code{radius} of \code{from} cells.
}
\details{
For each \code{from} cell, count the number of \code{to} cells within
\code{radius} microns. Report the number of \code{from} cells containing
at least \emph{one} \code{to} cell within \code{radius} as \code{from_with}.
Report the \emph{average} number of \code{to} cells as \code{within_mean}.

If \code{category} is specified, all reported values are for cells within
the given tissue category. If \code{category} is NULL, values are reported
for the entire data set.

\code{radius} may be a vector with multiple values.
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{subset_distance_matrix}}
}
