% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within_batch}
\alias{count_within_batch}
\title{Count cells within a radius for multiple tissue categories, phenotypes
and fields.}
\usage{
count_within_batch(base_path, pairs, radius, category = NA,
  phenotype_rules = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Path to a directory containing at least
one \code{_cell_seg_data.txt} file.}

\item{pairs}{A list of pairs of phenotypes. Each entry is a two-element
vector. The result will contain values for each pair.
A single pair can be passed as a plain vector.}

\item{radius}{The radius or radii to search within.}

\item{category}{Optional tissue categories to restrict both \code{from} and
\code{to} phenotypes.}

\item{phenotype_rules}{(Optional) A named list.
The item names are the phenotype names.
The values are selectors for \link{select_rows}.
For any phenotype not included in \code{phenotype_rules},
the name in \code{pairs} is used directly as the phenotype.}

\item{verbose}{If TRUE, display progress.}
}
\value{
A \code{data_frame} containing these columns:
\describe{
\item{\code{slide_id}}{Slide ID from the data files.}
\item{\code{source}}{Base file name of the source file with
\code{_cell_seg_data.txt} stripped off for brevity.}
\item{\code{category}}{Tissue category, if provided as a parameter,
or "all".}
\item{\code{from}}{From phenotype.}
\item{\code{to}}{To phenotype.}
\item{\code{radius}, \code{from_count}, \code{to_count},
\code{from_with}, \code{within_mean}}{Results from \link{count_within}
for this data file and tissue category.}
}
}
\description{
This is a batch version of \link{count_within}. Given the path to a directory
containing cell seg data files, for each given tissue category,
'from' phenotype, 'to' phenotype and radius, it counts the number of
\code{from} cells
having a \code{to} cell within \code{radius} microns.
}
\details{
\code{from} and \code{to} specify the phenotypes of interest.
Each one is a list of phenotype names.
Each list element is either a single name or a character vector.
If the element is a vector,
the phenotypes in the vector are treated collectively as a single type.
All pairs of \code{from} and \code{to} values are included in the output.

For example, the parameters
\code{from=list('cd68'), to=list('cd8', 'tumor'), radius=25}
would count \code{cd68} cells having a \code{cd8} cell within 25 microns and,
separately, \code{cd68} cells with a \code{tumor} cell within 25 microns.

The \code{category} parameter may be a single category or a list of categories.
}
\examples{
base_path = system.file("extdata", "TMA", package = "phenoptr")

# Count tumor cells near macrophages, and tumor cells near CD8 separately,
# in tumor and stroma tissue categories separately.
pairs = list(c('tumor', 'macrophage CD68'),
             c('tumor', 'cytotoxic CD8'))
radius = c(10, 25)
category = list('tumor', 'stroma')
count_within_batch(base_path, pairs, radius, category)

# Count tumor cells near any T cell in all tissue categories.
# Use `phenotype_rules` to define the T cell phenotype
pairs = c('tumor', 'T cell')
rules = list(
'T cell'=c('cytotoxic CD8', 'helper CD4', 'T reg Foxp3'))
count_within_batch(base_path, pairs, radius, phenotype_rules=rules)
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{count_within}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{subset_distance_matrix}}
}
