% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts_rtree.R
\name{count_within_detail}
\alias{count_within_detail}
\title{Compute count within for individual cells in a single field.}
\usage{
count_within_detail(csd, phenotypes = NULL, radii, whole_slide = FALSE)
}
\arguments{
\item{csd}{Cell seg data.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
will use \code{unique_phenotypes(csd)}. Counts are from each cell to each
phenotype.}

\item{radii}{The radius or radii to search within.}

\item{whole_slide}{If \code{TRUE}, allow multiple fields in \code{csd}. Otherwise
this is an error.}
}
\description{
Very fast version using \code{rtree::countWithinDistance()}.
}
