% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within_many_impl}
\alias{count_within_many_impl}
\title{Helper function for count_within_batch and count_within_many.
This does the actual work of calling count_within multiple times and
accumulating the result.}
\usage{
count_within_many_impl(csd, name, combos, radius, phenotype_rules)
}
\arguments{
\item{csd}{Cell seg data for a single field.}

\item{name}{Name associated with `csd`, for example the basename of the
image file.}

\item{combos}{List of pairs of (from phenotype name, to phenotype name)
and tissue category.}

\item{radius}{Vector of radii.}

\item{phenotype_rules}{Named list of phenotype rules.}
}
\description{
Helper function for count_within_batch and count_within_many.
This does the actual work of calling count_within multiple times and
accumulating the result.
}
