% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within_many_impl_dist}
\alias{count_within_many_impl_dist}
\title{Distance matrix implementation of count_within_many_impl}
\usage{
count_within_many_impl_dist(csd, name, combos, radius, phenotype_rules)
}
\description{
Distance matrix implementation of count_within_many_impl
}
\seealso{
count_within_many_impl
}
