% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts_rtree.R
\name{count_within_many_impl_rtree}
\alias{count_within_many_impl_rtree}
\title{rtree implementation of count_within_many_impl}
\usage{
count_within_many_impl_rtree(csd, name, combos, radii, phenotype_rules)
}
\description{
rtree implementation of count_within_many_impl
}
\seealso{
count_within_many_impl
}
