% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{create_buffer_bands}
\alias{create_buffer_bands}
\title{Create buffer regions around a boundary and within a region of interest}
\usage{
create_buffer_bands(boundary, roi, n, width)
}
\arguments{
\item{boundary}{Polygon for boundary line. Must be at least partially
within \code{roi}.}

\item{roi}{Polygon for ROI}

\item{n}{Number of buffers (on each side of boundary)}

\item{width}{Buffer width}
}
\value{
A list with two items
\itemize{
\item buffers - a data frame with index from -n to n and geometry column
\item divider - the dividing line for \code{boundary} within \code{roi}
}
}
\description{
For use with \code{density_by_region()}.
}
