% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_at_distance.R
\name{density_at_distance}
\alias{density_at_distance}
\title{Estimate cell density at distance from a tissue boundary.}
\usage{
density_at_distance(cell_seg_path, phenotypes, positive, negative,
  mask = TRUE, pixels_per_micron = getOption("phenoptr.pixels.per.micron"),
  smoother = "local", method = "reweight", ...)
}
\arguments{
\item{cell_seg_path}{Path to cell segmentation data.}

\item{phenotypes}{Optional named list of phenotypes to process.
\code{names(phenotypes)} are the names of the resulting phenotypes.
The values are in any format accepted by \code{\link{select_rows}}.
If omitted, will use all phenotypes in the cell seg data.}

\item{positive}{Name of the tissue category used as positive distance,
e.g. "stroma".}

\item{negative}{Name of the tissue category used as negative distance,
e.g. "tumor".}

\item{mask}{If true, compensate for edge effects by masking the image
to exclude cells which are closer to the edge of the image than to the
tissue boundary.}

\item{pixels_per_micron}{Conversion factor to microns.}

\item{smoother, method, ...}{Additional arguments passed to
\code{\link[spatstat]{rhohat}}.}
}
\value{
Returns a \code{list} containing two or three items:
\tabular{ll}{
\code{rho_hat} \tab The density estimate (see Details).\cr
\code{distance} \tab The distance map, a pixel image of class
\code{\link[spatstat]{im}}.\cr
\code{mask} \tab If \code{mask} is \code{TRUE},
a matrix showing the locations used.\cr
}
}
\description{
Given a cell seg table and an image containing masks for two tissue
classes, estimate the dependence of the density of cells of each
phenotype on the distance from the boundary between the two tissue classes.
}
\details{
The density estimate is computed by \code{\link[spatstat]{rhohat}}.
The signed distance from the boundary between the two tissue classes
is used as the covariate and density is estimated separately for each
phenotype.

The \code{rho_hat} element of the returned list is a
\code{\link[tibble]{data_frame}} containing the results of the density
estimation for each phenotype. It has six columns:
\tabular{ll}{
\code{phenotype} \tab The phenotype name.\cr
\code{distance} \tab Distance from tissue boundary (positive or negative),
i.e. the covariate for density estimation.\cr
\code{rho} \tab The density estimate from \code{\link[spatstat]{rhohat}},
in cells per square micron.\cr
\code{var}, \code{hi}, \code{lo} \tab The variance and high and low estimates of \code{rho}.
}
}
\section{Edge correction}{


The \code{\link[spatstat]{rhohat}} function does not have any built-in
edge correction. This may lead to incorrect density estimates
because it does not account for cells at the edge of the image which may
be near a tissue boundary which is not part of the image.

The \code{mask} parameter, if set to \code{TRUE}, restricts the density estimation
to cells which are closer to the tissue boundary in the image than they are
to the edge of the image or any other tissue class.
}

\examples{
\dontrun{
# Compute density for the sample data
values <- density_at_distance(sample_cell_seg_path(),
  list("CD8+", "CD68+", "FoxP3+"),
  positive="Stroma", negative="Tumor")

# Plot the densities
library(ggplot2)
ggplot(values[['rho_hat']], aes(distance, rho, color=phenotype)) +
  geom_line()

# Show the distance map with the mask superimposed and CK+ cells drawn
plot_diverging(values$distance, title='Distance')
plot(as.raster(values$mask), add=TRUE, col=c('black', 'transparent'))
}
}
\references{
A. Baddeley, E. Rubak and R.Turner.
Spatial Point Patterns: Methodology and Applications with R.
Chapman and Hall/CRC Press, 2015. Section 6.6.3.
}
