% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{density_bands}
\alias{density_bands}
\title{Estimate cell density in bands from a tissue boundary.}
\usage{
density_bands(cell_seg_path, phenotypes, positive, negative, width = 25,
  pixels_per_micron = getOption("phenoptr.pixels.per.micron"),
  map_path = NULL)
}
\arguments{
\item{cell_seg_path}{Path to a cell segmentation data file.}

\item{phenotypes}{Optional named list of phenotypes to process.
\code{names(phenotypes)} are the names of the resulting phenotypes.
The values are in any format accepted by \code{\link{select_rows}}.
If omitted, will use all phenotypes in the cell seg data.}

\item{positive}{Name of the tissue category used as positive distance,
e.g. "stroma".}

\item{negative}{Name of the tissue category used as negative distance,
e.g. "tumor".}

\item{width}{Width of the bands, in microns}

\item{pixels_per_micron}{Conversion factor to microns.}

\item{map_path}{Path to the segmentation map file. If NULL, look for the
map in the same directory as \code{cell_seg_path}.}
}
\value{
Returns a \code{list} with three items:
\tabular{ll}{
\code{densities} \tab A \code{data_frame} with five columns (see below).\cr
\code{cells} \tab Cell seg data with phenotypes updated per the \code{phenotypes}
parameter and an additional \code{distance} column.\cr
\code{distance} \tab The distance map, a pixel image
(\code{\link[spatstat]{im.object}}).\cr
}

The \code{densities} item contains five columns:
\tabular{ll}{
\code{phenotype} \tab The supplied phenotypes.\cr
\code{midpoint} \tab The midpoint of the distance band.\cr
\code{count} \tab The number of cells of the phenotype found
within the band.\cr
\code{area} \tab The area of the band, in square microns.\cr
\code{density} \tab The density of cells of the phenotype in the band,
in cells per square micron.\cr
}
}
\description{
Given a cell seg table and an image containing masks for two tissue
classes, estimate the density of cells of each specified
phenotype in bands from the boundary between the two
tissue classes.
}
\details{
\code{density_bands} uses a counting approach similar to a histogram.
First the image is divided into bands based on distance from the
specified boundary. Next, the number of cells of each phenotype
within each distance band is counted and the area of each band
is estimated. The density estimates are the ratios of the cell
counts to the area estimates.

Density estimates are
in cells per square micron; multiply by 1,000,000 for cells per square
millimeter.

The returned value includes the cell counts and area of each band,
making it straightforward to aggregate across multiple fields from a
single sample. The aggregate density is computed by summing the
cell counts and areas across all fields from a sample, then dividing
to compute density.
}
\examples{
# Compute density for the sample data
values <- density_bands(sample_cell_seg_path(),
  list("CD8+", "CD68+", "FoxP3+"),
  positive="Stroma", negative="Tumor")

# Plot the densities in a single plot
library(ggplot2)
ggplot(values$densities, aes(midpoint, density*1000000, color=phenotype)) +
  geom_line(size=2) +
  labs(x='Distance from tumor boundary (microns)',
       y='Estimated cell density (cells per sq mm)')
}
\seealso{
Other density estimation: \code{\link{density_at_distance}}
}
\concept{density estimation}
