% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{density_by_region}
\alias{density_by_region}
\title{Compute density of each phenotype in each given region}
\usage{
density_by_region(csd, regions, phenotypes = NULL)
}
\arguments{
\item{csd}{Cell seg table}

\item{regions}{An \code{sf::st_sf} or \code{sf::st_sfc} object containing the regions
of interest, for example the output of \code{create_buffers()}.}

\item{phenotypes}{Phenotypes of interest; if omitted, will use
\code{phenoptr::unique_phenotypes(csd)}.}
}
\value{
A data frame, \code{regions} annotated with an area column
(in square microns) and two columns per phenotype
giving the count and density of that phenotype in that region.
}
\description{
Compute density of each phenotype in each given region
}
