% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Create a distance matrix from cell seg data.}
\usage{
distance_matrix(d)
}
\arguments{
\item{d}{A data frame with \code{Cell X Position} and \code{Cell Y Position}
columns.}
}
\value{
A square matrix with both dimensions equal to \code{nrow(d)}
}
\description{
Create a distance matrix from cell seg data.
}

