% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{find_nearest_distance_dist}
\alias{find_nearest_distance_dist}
\title{Distance-matrix implementation of \code{find_nearest_distance}.}
\usage{
find_nearest_distance_dist(csd, phenotypes = NULL, dst = NULL)
}
\description{
Distance-matrix implementation of \code{find_nearest_distance}.
}
\seealso{
find_nearest_distance
}
