% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{find_nearest_distance_rtree}
\alias{find_nearest_distance_rtree}
\title{\code{rtree}-based implementation of \code{find_nearest_distance}.}
\usage{
find_nearest_distance_rtree(csd, phenotypes = NULL)
}
\description{
\code{rtree}-based implementation of \code{find_nearest_distance}.
}
\seealso{
find_nearest_distance
}
