% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_invert.R
\docType{data}
\name{geom_sf_invert}
\alias{geom_sf_invert}
\alias{StatSfInvert}
\alias{stat_sf_invert}
\alias{scale_sf_invert}
\title{Negate the y-axis of a simple features object}
\usage{
geom_sf_invert(...)

scale_sf_invert(...)
}
\arguments{
\item{...}{Passed on to \code{ggplot2::geom_sf()} or \code{ggplot2::scale_y_continuous()}}
}
\description{
Use with \code{scale_sf_invert()} to match the orientation of Polaris images.
}
\keyword{datasets}
