% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{get_map_path}
\alias{get_map_path}
\title{Get the path to the segmentation map file for a field}
\usage{
get_map_path(field_name, export_path)
}
\arguments{
\item{field_name}{Name of the field of interest}

\item{export_path}{Path to the inForm export directory}
}
\value{
The path to the \code{binary_seg_maps} file for the given field.
}
\description{
Get the path to the segmentation map file for a field
}
