% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{get_tissue_category_index}
\alias{get_tissue_category_index}
\title{Get the mapping from tissue category name to label number for a
tissue category map.}
\usage{
get_tissue_category_index(map)
}
\arguments{
\item{map}{A tissue category image from read_maps.}
}
\value{
A named vector mapping tissue category names to label numbers
for the map.
}
\description{
Get the mapping from tissue category name to label number for a
tissue category map.
}
