% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informr.R
\docType{package}
\name{informr-package}
\alias{informr}
\alias{informr-package}
\title{Helpers for working with inForm data}
\description{
informr contains functions that make it easier to read and analyze data tables
and images created by PerkinElmer's inForm software.
}
\details{
informr is part of the PerkinElmer Phenoptics family of
Quantitative Pathology Research Solutions. For more information visit the
Phenoptics \href{http://www.perkinelmer.com/cancer-immunology/index.html}{homepage}.
}
\section{Package options}{

\code{read_cell_seg_data} converts pixel measurements to microns. Several other
functions also implicitly convert pixels to microns. The default conversion
is given by \code{getOption('informr.pixels.per.micron')}, which has a default
value of 2 pixels/μm, i.e. pixels are 0.5 μm square. To use a different
value, either pass a \code{pixels_per_micron} parameter to functions which take
one, or set \code{options(informr.pixels.per.micron=<new value>)}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PerkinElmer/informr}
  \item Report bugs at \url{https://github.com/PerkinElmer/informr/issues}
}

}
\author{
\strong{Maintainer}: Kent Johnson \email{kent.johnson@perkinelmer.com}

Other contributors:
\itemize{
  \item PerkinElmer, Inc. [copyright holder]
}

}
