% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cell_seg_data.R
\name{list_cell_seg_files}
\alias{list_cell_seg_files}
\title{Find inForm data files.}
\usage{
list_cell_seg_files(path, ...)
}
\arguments{
\item{path}{Path to the base directory to search.}

\item{...}{Additional arguments passed to \link[base:list.files]{list.files}.
Pass \code{recursive=TRUE} to search a directory hierarchy.}
}
\value{
A list of  file paths.
}
\description{
\code{list_cell_seg_files} finds inForm cell seg data files in a single directory
or a directory hierarchy.
}
\seealso{
Other file readers: \code{\link{read_cell_seg_data}},
  \code{\link{read_maps}}
}
