% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{make_phenotype_column}
\alias{make_phenotype_column}
\title{Mutate a cell seg table to have a Phenotype column with the
desired phenotypes.}
\usage{
make_phenotype_column(csd, phenotypes = NULL)
}
\arguments{
\item{csd}{A cell seg table.}

\item{phenotypes}{A named vector or list of phenotype rules. If NULL, use
`unique_phenotypes(csd)`.}
}
\value{
A new cell seg table
}
\description{
Note: Cells that satisfy multiple phenotype definitions will appear
multiple times in the result.
}
