% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{make_ppp}
\alias{make_ppp}
\title{Make a marked point pattern for cells in a single field}
\usage{
make_ppp(csd, export_path, pheno, field_name = NULL, tissue_categories = NULL)
}
\arguments{
\item{csd}{Cell seg data, may contain multiple fields}

\item{export_path}{Path to a directory containing composite and component
image files from inForm}

\item{pheno}{Phenotype definition. Either a (possibly named) character
vector to be parsed by \code{\link[=parse_phenotypes]{parse_phenotypes()}} or a named list as
used by \code{\link[=select_rows]{select_rows()}}.}

\item{field_name}{Sample Name or Annotation ID for the field of interest.
May be omitted if \code{csd} contains data for only one field.}

\item{tissue_categories}{Tissue categories of interest. If supplied, the
returned \code{ppp} will contain only cells in these categories and the
associated window will be restricted to the extent of these categories.}
}
\value{
Returns a marked point pattern (\code{\link[spatstat:ppp]{spatstat::ppp}} object)
with a single mark value.
}
\description{
Makes a marked point pattern (\code{\link[spatstat:ppp]{spatstat::ppp}}) for the cells
of the specified phenotype in the
specified tissue categories and field.
}
\examples{
# ppp for CD8+ cells in the sample data
suppressPackageStartupMessages(library(spatstat))
pp <- make_ppp(sample_cell_seg_data, sample_cell_seg_folder(),
  "CD8+", tissue_categories="Tumor")
plot(pp, show.window=FALSE, main='')

# To include multiple phenotypes in a single point pattern,
# create them separately and join them with [spatstat::superimpose()].

pp2 <- make_ppp(sample_cell_seg_data, sample_cell_seg_folder(),
  "CK+", tissue_categories="Tumor")
plot(superimpose(pp2, pp), cols=c('red', 'blue'), show.window=FALSE, main='')
}
