% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{map_as_raster}
\alias{map_as_raster}
\title{Make a spatially-referenced raster from a map image}
\usage{
map_as_raster(map)
}
\arguments{
\item{map}{Map image from \code{read_maps}}
}
\description{
Make a spatially-referenced raster from a map image
}
