% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_composites.R
\name{parse_composite_info}
\alias{parse_composite_info}
\title{Parse the ImageDescription of a composite image.}
\usage{
parse_composite_info(info)
}
\arguments{
\item{info}{The ImageDescription string}
}
\value{
A list, containing none, some or all of these items:
\itemize{
\item composite_name - the user's name for this composite
\item components - a named vector, where names are component names and values
are hex colors
\item scheme_name - the user's name for the phenotype scheme
\item phenotypes -  a named vector, where names are phenotype names and values
are hex colors
}
}
\description{
Parse the ImageDescription of a composite image.
}
\keyword{internal}
