% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_composites.R
\name{parse_names_and_colors}
\alias{parse_names_and_colors}
\title{Parse the Entry items in a Composite or Phenotypes node}
\usage{
parse_names_and_colors(node, name_node_name)
}
\arguments{
\item{node}{The XML node to parse}

\item{name_node_name}{Name of the node containing item names}
}
\value{
A named vector of colors
}
\description{
Parse the Entry items in a Composite or Phenotypes node
}
\keyword{internal}
