% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{parse_rect}
\alias{parse_rect}
\title{Parse a single rectangle annotation}
\usage{
parse_rect(field)
}
\arguments{
\item{field}{Parsed XML for a single rectangle annotation}
}
\value{
An \code{sf::st_polygon} with coordinates of the rectangle
}
\description{
Parse a single rectangle annotation
}
\keyword{internal}
