% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{parse_roi_rects}
\alias{parse_roi_rects}
\title{Parse the rectangles from an ROI annotation}
\usage{
parse_roi_rects(roi)
}
\arguments{
\item{roi}{Parsed XML for an ROI annotation}
}
\value{
An \code{sf::sfc} object with geometries for each rectangle in roi
}
\description{
Parse the rectangles from an ROI annotation
}
\keyword{internal}
