% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{parse_tags}
\alias{parse_tags}
\title{Parse tags from a node of an annotations file}
\usage{
parse_tags(node, ns = xml2::xml_ns(node))
}
\arguments{
\item{node}{XML node containing tags}

\item{ns}{Optional namespace; providing this makes parsing faster.}
}
\value{
A string with space-separated tags
}
\description{
Parse tags from a node of an annotations file
}
\keyword{internal}
