% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{parse_tissue_description}
\alias{parse_tissue_description}
\title{Parse the description tag of the tissue map image read from an
inForm \code{binary_seg_maps} file}
\usage{
parse_tissue_description(img)
}
\arguments{
\item{img}{The \code{Tissue} image from \code{\link{read_maps}}.}
}
\value{
A named vector whose names are the tissue classes in \code{img} and
whose values are the mask values for the class in \code{img}.
}
\description{
Parse the description tag of the tissue map image read from an
inForm \code{binary_seg_maps} file
}
