% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoptr.R
\docType{package}
\name{phenoptr-package}
\alias{phenoptr}
\alias{phenoptr-package}
\title{Helpers for working with inForm data}
\description{
phenoptr contains functions that make it easier to read and analyze data
tables and images created by Akoya Biosciences' inForm software.
}
\details{
phenoptr is part of the Akoya Biosciences Phenoptics family of
Quantitative Pathology Research Solutions. For more information visit the
Phenoptics \href{http://www.perkinelmer.com/cancer-immunology/index.html}{home page}.
}
\section{Package options}{

\code{read_cell_seg_data} converts pixel measurements to microns. Several other
functions also implicitly convert pixels to microns. The default conversion
is given by \code{getOption('phenoptr.pixels.per.micron')}, which has a default
value of 2 pixels/micron, i.e. pixels are 0.5 micron square
(the resolution of 20x MSI fields taken on Vectra Polaris and Vectra 3).
To use a
different value, either pass a \code{pixels_per_micron} parameter to functions
which take one, or set \code{options(phenoptr.pixels.per.micron=<new value>)}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://akoyabio.github.io/phenoptr/}
  \item Report bugs at \url{https://github.com/akoyabio/phenoptr/issues}
}

}
\author{
\strong{Maintainer}: Kent S Johnson \email{kjohnson@akoyabio.com}

Other contributors:
\itemize{
  \item Akoya Biosciences [copyright holder]
}

}
