% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{phenotype_columns}
\alias{phenotype_columns}
\title{Find the columns used by phenotype formulae}
\usage{
phenotype_columns(phenos)
}
\arguments{
\item{phenos}{A list of parsed phenotypes}
}
\value{
A vector of names or NULL if none found
}
\description{
Given a list of parsed phenotypes, e.g. from \code{parse_phenotypes},
return a vector containing all the names referenced
by formulae in \code{phenos}.
}
