% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{phenotype_names}
\alias{phenotype_names}
\title{Make user-friendly names for phenotypes}
\usage{
phenotype_names(phenos)
}
\arguments{
\item{phenos}{A list or vector of phenotype definitions
as used by `parse_phenotypes`.}
}
\value{
A vector of name for `phenos`.
}
\description{
Make user-friendly names for phenotypes
}
\keyword{internal}
