% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{plot_buffers}
\alias{plot_buffers}
\title{Plot density bands with divider and ROI}
\usage{
plot_buffers(buffers, divider, roi)
}
\arguments{
\item{buffers}{The band buffers}

\item{divider}{The tumor boundary line}

\item{roi}{The region of interest boundary}
}
\value{
A \code{ggplot} object
}
\description{
Plot density bands with divider and ROI
}
