% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{plot_density_by_band}
\alias{plot_density_by_band}
\title{Plot density by band}
\usage{
plot_density_by_band(densities, colors)
}
\arguments{
\item{densities}{Result of calling \code{density_by_region()}}

\item{colors}{Named vector of phenotype colors}
}
\value{
A \code{ggplot} object
}
\description{
Plot density by band
}
