% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_at_distance.R
\name{plot_diverging}
\alias{plot_diverging}
\title{Plot a distance map using a diverging color scale with white (grey) at 0}
\usage{
plot_diverging(im, title)
}
\arguments{
\item{im}{A pixel image of class \code{\link[spatstat]{im}}.}

\item{title}{Title for the plot.}
}
\description{
Plot a distance map using a diverging color scale with white (grey) at 0
}
