% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_components.R
\name{read_components}
\alias{read_components}
\title{Read a component TIFF file.}
\usage{
read_components(path)
}
\arguments{
\item{path}{Path to the \code{component_data.tif} file.}
}
\value{
A named list of matrices, one for each component.
}
\description{
Read an inForm \code{component_data.tif} file.
If the file contains multiple resolutions or other data,
keep only the full-resolution component images,
extract the component names from the image descriptions,
and return a named list of image matrices.
}
\details{
Images are oriented to match the coordinates in a cell seg data file,
i.e. (0, 0) at the top left and the row corresponding to Y and
column corresponding to X.

Limited to 1x1 fields due to
limitations in the \code{tiff} package.
}
\examples{
\dontrun{
path = "/path/to/my_component_data.tif"
components = read_components(path)
}
}
\seealso{
Other file readers: \code{\link{list_cell_seg_files}},
  \code{\link{read_cell_seg_data}}, \code{\link{read_maps}}
}
