% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_composites.R
\name{read_composite_info}
\alias{read_composite_info}
\title{Read just the composite info from images in a TIFF composite image file.}
\usage{
read_composite_info(path)
}
\arguments{
\item{path}{Path to a TIFF composite}
}
\value{
A list of lists containing some of
\itemize{
\item composite_name - the user's name for this composite
\item components - a named vector, where names are component names and values
are hex colors
\item scheme_name - the user's name for the phenotype scheme
\item phenotypes -  a named vector, where names are phenotype names and values
are hex colors
}
}
\description{
Read just the composite info from images in a TIFF composite image file.
}
