% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_composites.R
\name{read_composites}
\alias{read_composites}
\title{Read composite images}
\usage{
read_composites(path)
}
\arguments{
\item{path}{Path to the composite image file}
}
\value{
A list of image matrices, possibly containing image metadata
in attributes (see Details).
}
\description{
Read images based on inForm composite, including composite with
phenotype overlay, and phenotype overlay alone.
}
\details{
If the selected file is a JPEG file, the returned value will be a list
containing a single image matrix and no attributes other than \code{dim}.

If the selected file is a TIFF file, the returned value will be a list
possibly containing multiple image matrices with many metadata attributes.
In addition to the TIFF metadata provided by \code{tiff::readTIFF()},
attributes may include
\itemize{
\item composite_name - the user's name for this composite
\item components - a named list, where names are component names and values
are hex colors
\item scheme_name - the user's name for the phenotype scheme
\item phenotypes -  a named list, where names are phenotype names and values
are hex colors
}
}
