% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_components.R
\name{read_field_info}
\alias{read_field_info}
\title{Read information about a field from a component image file.}
\usage{
read_field_info(field_name, export_path)
}
\arguments{
\item{field_name}{Name of the field}

\item{export_path}{Directory containing the component data file}
}
\value{
The result of calling \code{\link[=get_field_info]{get_field_info()}} on the correct
component file.
}
\description{
Find the location, size and magnification of an inForm field by inspecting a
\code{component_data.tif} file.
}
