% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{read_tagged_rois}
\alias{read_tagged_rois}
\title{Read and combine all tagged ROIS from an annotation file}
\usage{
read_tagged_rois(annotation_file)
}
\arguments{
\item{annotation_file}{Path to annotations file, either XML or GeoJSON}
}
\value{
A named list of (multi)polygons, one for each tag
in the annotations file
}
\description{
Reads all tagged ROIs from the provided file. ROIs with the same
tag are combined with \code{sf::st_union()}. A single (multi)polygon
object is returned for each unique tag.
}
