% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cell_seg_data}
\alias{sample_cell_seg_data}
\title{Sample cell segmentation data.}
\format{A data frame with 6072 rows and 199 variables}
\usage{
sample_cell_seg_data
}
\description{
A dataset containing inForm cell segmentation data for a single field.
This is the data from\preformatted{system.file("extdata", "sample",
                  "Set4_1-6plex_[16142,55840]_cell_seg_data.txt",
                  package = "phenoptr")
}
}
\details{
This table shows the Opal stains used, the epitope they were bound to,
and the colors used to show the stain in the composite and phenotype views.

\tabular{llll}{
Stain   \tab Epitope \tab Composite \tab Phenotype \cr
----------- \tab ----------- \tab ------------- \tab ------------- \cr
DAPI    \tab Nucleus \tab Blue    \tab N/A    \cr
Opal520 \tab PDL1    \tab Red     \tab N/A    \cr
Opal540 \tab CD8     \tab Yellow  \tab Yellow \cr
Opal570 \tab FoxP3   \tab Orange  \tab Orange \cr
Opal620 \tab CD68    \tab Magenta \tab Magenta\cr
Opal650 \tab PD1     \tab Green   \tab N/A    \cr
Opal690 \tab CK      \tab Cyan    \tab Cyan
}
The sample cell data includes phenotypes CD8+ (cytotoxic T cell),
CD68+ (macrophage),
FoxP3+ (regulatory T cell), CK+ (tumor), and other.

To see all the files included for this field, use
\code{list.files(system.file("extdata", "sample", package = "phenoptr"))}.

For more sample data from this and related samples, see the
\href{https://perkinelmer.github.io/phenoptrExamples}{phenoptrExamples} package.
}
\keyword{datasets}
