% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{select_rows}
\alias{select_rows}
\title{Flexibly select rows of a data frame.}
\usage{
select_rows(csd, sel, phenotype_column = "Phenotype")
}
\arguments{
\item{csd}{A data frame}

\item{sel}{May be a character vector, a one-sided formula or a list
containing such. A character vector is interpreted as the name(s) of one or
more phenotypes and selects any matching phenotype. A formula is
interpreted as an expression on the columns of \code{csd}. Multiple items are
joined with AND.}

\item{phenotype_column}{Optional, name of the phenotype column to use for
simple selection}
}
\value{
A logical vector of length \code{nrow(csd)} which selects rows
  according to sel.
}
\description{
Select rows of a data frame based on phenotypes or other
expressions.
}
\examples{
csd = sample_cell_seg_data

# Select tumor cells with PDL1 expression > 4
selector = list('tumor', ~`Entire Cell PDL1 (Opal 620) Mean`>4)
pdl1_pos_tumor = csd[select_rows(csd, selector),]
range(pdl1_pos_tumor$`Entire Cell PDL1 (Opal 620) Mean`)

# Select all T-cells. Note: Use c() to combine phenotypes, not list()
selector = c('cytotoxic CD8', 'helper CD4', 'T reg Foxp3')
tcells = csd[select_rows(csd, selector),]
table(tcells$Phenotype)
}
