% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distribution_report.R
\name{spatial_distribution_report}
\alias{spatial_distribution_report}
\title{Create a spatial distribution report}
\usage{
spatial_distribution_report(cell_seg_path, phenotypes, colors,
  plot_gx_max = 100, micron_line = 25)
}
\arguments{
\item{phenotypes}{A list of phenotype specifications in the form used by
\code{\link[=select_rows]{select_rows()}}.}

\item{colors}{A character vector of color specifications, of the same
length as \code{phenotypes}.}

\item{plot_gx_max}{Optional, maximum value of radius to show on G
function plots.}

\item{micron_line}{Optional, X location (radius in microns)
of crosshair on G function plots.}

\item{path}{Path to a cell seg table.}
}
\description{
Creates an HTML report showing the location and nearest-neighbor relations
of cells in a single field and two chosen phenotypes.
}
\details{
A cell seg data file for the field is required. If an image of the field
is available, it will be used as a background.

The report is saved to the same directory as the input file.
}
