% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{split_and_trim}
\alias{split_and_trim}
\title{Split a single string and trim whitespace from the results}
\usage{
split_and_trim(str, pattern)
}
\arguments{
\item{str}{A single string.}

\item{pattern}{Pattern to split on.}
}
\value{
A character vector of split components.
}
\description{
Split a single string and trim whitespace from the results
}
\keyword{internal}
