% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{split_quoted}
\alias{split_quoted}
\title{Split a single string on / and trim white space from the results.
Slashes quoted with backquotes are not split.
\code{str} must not contain tabs}
\usage{
split_quoted(str, split = "/")
}
\arguments{
\item{str}{A single string.}

\item{split}{Character vector containing characters to split on}
}
\value{
A character vector of split components.
}
\description{
Split a single string on / and trim white space from the results.
Slashes quoted with backquotes are not split.
\code{str} must not contain tabs
}
\keyword{internal}
