% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stop_if_multiple_fields}
\alias{stop_if_multiple_fields}
\title{Stop if \code{csd} contains multiple fields}
\usage{
stop_if_multiple_fields(csd)
}
\description{
Stop if \code{csd} contains multiple fields
}
\keyword{internal}
