% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{trim_tissue_categories}
\alias{trim_tissue_categories}
\title{Compute corrected tissue category areas per annotation by clipping
to a provided polygon and counting pixels.}
\usage{
trim_tissue_categories(annotations, roi, export_path, plot_path = NULL)
}
\arguments{
\item{annotations}{Names of annotations to process}

\item{roi}{Clipping polygon, as an \code{sf::st_sf} or \code{sf::st_sfc} object.}

\item{export_path}{Path to a directory containing binary_seg_map files.
These files must include TIFF tags for location and resolution.}

\item{plot_path}{If not \code{NULL}, the file path at which to save
a check plot showing the given ROI and the trimmed tissue categories.}
}
\value{
A data frame containing computed tissue category areas for
all annotations
}
\description{
Compute corrected tissue category areas per annotation by clipping
to a provided polygon and counting pixels.
}
