% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{unique_phenotypes}
\alias{unique_phenotypes}
\title{Find unique phenotypes in a cell seg table}
\usage{
unique_phenotypes(csd)
}
\arguments{
\item{csd}{A cell seg table such as read by `read_cell_seg_table`.}
}
\value{
A named character vector containing the phenotype names.
}
\description{
For cell seg tables containing a single `Phenotype` column, this
returns a vector containing all the non-blank phenotypes in the table.
For cell seg tables containing multiple phenotype columns, it returns
a vector with just the positive phenotypes.
}
