% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{validate_phenotype_definitions}
\alias{validate_phenotype_definitions}
\title{Validate a user-specified phenotype definition}
\usage{
validate_phenotype_definitions(pheno, available, csd = NULL)
}
\arguments{
\item{pheno}{Text description of a phenotype,
for `phenoptr::parse_phenotypes`.}

\item{available}{A character vector of available phenotypes}

\item{csd}{If supplied, any formula arguments will be checked for
validity against `csd`.}
}
\value{
An error message or empty string
}
\description{
Check that the specified phenotype
can be formed from available phenotypes and data columns.
}
