% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{validate_phenotypes}
\alias{validate_phenotypes}
\title{Validate a phenotype parameter}
\usage{
validate_phenotypes(phenotypes, csd)
}
\arguments{
\item{phenotypes}{Nominally, a list or vector of phenotype definitions,
or NULL.}

\item{csd}{A cell seg table}
}
\value{
A named list of phenotype definitions
}
\description{
Validate a phenotype parameter
}
