% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{compute_all_nearest_distance}
\alias{compute_all_nearest_distance}
\title{Nearest neighbors from a file.}
\usage{
compute_all_nearest_distance(cell_table_path = NULL, out_path = NULL)
}
\arguments{
\item{cell_table_path}{Path to an inForm cell seg data file, or NULL
to prompt for the path.}

\item{out_path}{Path to the output file, or NULL to create a path from the
input file path.}
}
\description{
Compute nearest distance to each phenotype for each cell in a
(possibly merged) inForm cell seg table. Add \code{Distance to <phenotype>}
columns.
Write the result to a new file.
}
\details{
NOTE: The input file is read using \link{read_cell_seg_data} so the conversions
and cleanup it does will be applied to the output data.
}
\seealso{
\link{find_nearest_distance} which performs the distance calculation.

Other distance functions: \code{\link{count_touching_cells}},
  \code{\link{count_within_batch}},
  \code{\link{count_within}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{spatial_distribution_report}},
  \code{\link{subset_distance_matrix}}
}
