% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within}
\alias{count_within}
\title{Find the number of cells within a radius}
\usage{
count_within(csd, dst, from, to, radius, category = NULL)
}
\arguments{
\item{csd}{A data frame with \code{Cell X Position},
\code{Cell Y Position} and \code{Phenotype} columns,
such as the result of calling \code{\link{read_cell_seg_data}}.}

\item{dst}{The distance matrix corresponding to \code{csd},
produced by calling \code{\link{distance_matrix}}.}

\item{from, to}{Selection criteria for the
rows and columns. Accepts all formats accepted by \code{\link{select_rows}}.}

\item{radius}{The radius to search within.}

\item{category}{Optional tissue category to restrict both \code{from} and
\code{to}.}
}
\value{
A \code{\link{data_frame}} with four columns:
  \describe{
   \item{\code{from_count}}{The number of \code{from} cells found.}
   \item{\code{to_count}}{The number of \code{to} cells found.}
   \item{\code{within_count}}{The number of \code{to} cells found within
   \code{radius} microns of a \code{from} cell.}
   \item{\code{within_mean}}{The average number of \code{to} cells found
   within \code{radius} microns of a \code{from} cell.}
 }
}
\description{
Find the number of cells of type \code{to} within \code{radius}
of \code{from} cells
within tissue category \code{category}.
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{subset_distance_matrix}}
}
