% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within}
\alias{count_within}
\title{Count cells within a radius for a single field.}
\usage{
count_within(csd, from, to, radius, category = NA, dst = NULL)
}
\arguments{
\item{csd}{A data frame with \code{Cell X Position},
\code{Cell Y Position} and \code{Phenotype} columns,
such as the result of calling \code{\link{read_cell_seg_data}}.}

\item{from, to}{Selection criteria for the
rows and columns. Accepts all formats accepted by \code{\link{select_rows}}.}

\item{radius}{The radius or radii to search within.}

\item{category}{Optional tissue category to restrict both \code{from} and
\code{to}.}

\item{dst}{Optional distance matrix corresponding to \code{csd},
produced by calling \code{\link{distance_matrix}}.}
}
\value{
A \code{\link{tibble}} with five columns and one row for each
value in \code{radius}:
\describe{
\item{\code{radius}}{The value of \code{radius} for this row.}
\item{\code{from_count}}{The number of \code{from} cells found in
\code{csd}.}
\item{\code{to_count}}{The number of \code{to} cells found in \code{csd}.}
\item{\code{from_with}}{The number of \code{from} cells with a
\code{to} cell within \code{radius}.}
\item{\code{within_mean}}{The average number of \code{to} cells found
within \code{radius} microns of each \code{from} cell.}
}
}
\description{
Count the number of \code{from} cells having a \code{to} cell within
\code{radius} microns in tissue category \code{category}.
Compute the average number of \code{to} cells
within \code{radius} of \code{from} cells.
}
\details{
For each \code{from} cell, count the number of \code{to} cells within
\code{radius} microns. Report the number of \code{from} cells containing
at least \emph{one} \code{to} cell within \code{radius} as \code{from_with}.
Report the \emph{average} number of \code{to} cells per
\code{from} cell as \code{within_mean}.

\code{count_within} counts cells within a single field. It will give an
error if run on a merged cell seg data file. To count cells in a merged file,
use \code{\link[dplyr]{group_by}} and \code{\link[dplyr]{do}} to call
\code{count_within} for each sample in the merged file. See the Examples.

There are some subtleties to the count calculation.
\itemize{
\item It is not symmetric in \code{from} and \code{to}.
For example the number of tumor cells with a
macrophage within 25 microns is not the same as the number of macrophages
with a tumor cell within 25 microns.
\item \code{from_count*within_mean} is \emph{not} the number of
\code{to} cells within \code{radius} of a \code{from} cell, it may
count \code{to} cells multiple times.
\item Surprisingly, \code{from_count*within_mean} is symmetric in
\code{from} and \code{to}. The double-counting works out.
}

To aggregate \code{within_mean} across multiple samples (e.g. by Slide ID)
see the examples below.

If \code{category} is specified, all reported values are for cells within
the given tissue category. If \code{category} is NA, values are reported
for the entire data set.

\code{radius} may be a vector with multiple values.
}
\examples{
library(dplyr)
csd <- sample_cell_seg_data

# Find the number of macrophages with a tumor cell within 10 or 25 microns
count_within(csd, from='CD68+', to='CK+', radius=c(10, 25))

# Find the number of tumor cells with a macrophage within 10 or 25 microns
count_within(csd, from='CK+', to='CD68+', radius=c(10, 25))

\dontrun{
# If 'merged' is a merged cell seg file, this will run count_within for
# each field:
distances = merged \%>\% group_by(`Slide ID`, `Sample Name`) \%>\%
  do(count_within(., from='CK+', to='CD68+', radius=c(10, 25)))

# This will aggregate the fields by Slide ID:
distances \%>\% group_by(`Slide ID`, radius) \%>\%
  summarize(within=sum(from_count*within_mean, na.rm=TRUE),
            from_count=sum(from_count),
            to_count=sum(to_count),
            from_with=sum(from_with),
            within_mean=within/from_count) \%>\%
  select(-within)
}
}
\seealso{
Other distance functions: 
\code{\link{compute_all_nearest_distance}()},
\code{\link{count_touching_cells}()},
\code{\link{count_within_batch}()},
\code{\link{count_within_many}()},
\code{\link{distance_matrix}()},
\code{\link{find_nearest_distance}()},
\code{\link{spatial_distribution_report}()},
\code{\link{subset_distance_matrix}()}
}
\concept{distance functions}
