% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts.R
\name{count_within_batch}
\alias{count_within_batch}
\title{Count cells within a radius for multiple tissue categories, phenotypes
and fields.}
\usage{
count_within_batch(base_path, pairs, radius, category = NA,
  phenotype_rules = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Path to a directory containing at least
one \code{_cell_seg_data.txt} file.}

\item{pairs}{A list of pairs of phenotypes. Each entry is a two-element
vector. The result will contain values for each pair.}

\item{radius}{The radius or radii to search within.}

\item{category}{Optional tissue categories to restrict both \code{from} and
\code{to} phenotypes.}

\item{phenotype_rules}{(Optional) A named list.
Item names are phenotype names and must match entries in \code{pairs}.
Item values are selectors for \link{select_rows}.}

\item{verbose}{If TRUE, display progress.}
}
\value{
A \code{data_frame} containing these columns:
\describe{
\item{\code{slide_id}}{Slide ID from the data files.}
\item{\code{source}}{Base file name of the source file with
\code{_cell_seg_data.txt} stripped off for brevity.}
\item{\code{category}}{Tissue category, if provided as a parameter,
or "all".}
\item{\code{from}}{From phenotype.}
\item{\code{to}}{To phenotype.}
\item{\code{radius}, \code{from_count}, \code{to_count},
\code{from_with}, \code{within_mean}}{Results from \link{count_within}
for this data file and tissue category.}
}
}
\description{
This is a batch version of \link{count_within}. Given the path to a directory
containing cell seg data files, for each given tissue category,
pair of
'from' phenotype and 'to' phenotype, and radius, it counts the number of
'from' cells
having a 'to' cell within \code{radius} microns.
}
\details{
The \code{category} parameter may be a single category or a list of categories.

See the vignette
\emph{Selecting cells within a cell segmentation table} for more on
the use of \code{pairs} and \code{phenotype_rules}.
}
\examples{
base_path = sample_cell_seg_folder()

# Count tumor cells near macrophages, and tumor cells near CD8 separately,
# in tumor and stroma tissue categories separately.
pairs = list(c('CK+', 'CD68+'),
             c('CK+', 'CD8+'))
radius = c(10, 25)
category = list('Tumor', 'Stroma')
count_within_batch(base_path, pairs, radius, category)

# Count tumor cells near any T cell in all tissue categories.
# Use `phenotype_rules` to define the T cell phenotype
pairs = c('CK+', 'T cell')
rules = list(
'T cell'=c('CD8+', 'FoxP3+'))
count_within_batch(base_path, pairs, radius, phenotype_rules=rules)
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{count_touching_cells}},
  \code{\link{count_within}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{spatial_distribution_report}},
  \code{\link{subset_distance_matrix}}
}
