% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_counts_rtree.R
\name{count_within_many_impl_rtree}
\alias{count_within_many_impl_rtree}
\title{rtree implementation of count_within_many_impl}
\usage{
count_within_many_impl_rtree(csd, name, combos, radii, phenotype_rules)
}
\arguments{
\item{csd}{Cell seg data for a single field.}

\item{name}{Name associated with \code{csd}, for example the basename of the
image file.}

\item{combos}{List of pairs of (from phenotype name, to phenotype name)
and tissue category.}

\item{radii}{Vector of radii.}

\item{phenotype_rules}{Named list of phenotype rules.}
}
\description{
rtree implementation of count_within_many_impl
}
\seealso{
count_within_many_impl
}
