% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_at_distance.R
\name{density_at_distance}
\alias{density_at_distance}
\title{Estimate cell density at distance from a tissue boundary.}
\usage{
density_at_distance(cell_seg_path, phenotypes, positive, negative,
  mask = FALSE, pixels_per_micron = getOption("phenoptr.pixels.per.micron"),
  ...)
}
\arguments{
\item{cell_seg_path}{Path to a cell segmentation data file.}

\item{phenotypes}{Optional named list of phenotypes to process.
\code{names(phenotypes)} are the names of the resulting phenotypes.
The values are in any format accepted by \code{\link{select_rows}}.
If omitted, will use all phenotypes in the cell seg data.}

\item{positive}{Name of the tissue category used as positive distance,
e.g. "stroma".}

\item{negative}{Name of the tissue category used as negative distance,
e.g. "tumor".}

\item{mask}{If true, compensate for edge effects by masking the image
to exclude cells which are closer to the edge of the image than to the
tissue boundary.}

\item{pixels_per_micron}{Conversion factor to microns.}

\item{...}{Additional arguments passed to \code{\link[spatstat]{rhohat}}.
Default parameters are \code{method="ratio", smoother="kernel", bw="nrd"}.}
}
\value{
Returns a \code{list} containing four items:
\tabular{ll}{
\code{points} \tab The points used, marked with their phenotype,
a \code{\link[spatstat]{ppp.object}}.\cr
\code{rhohat} \tab The density estimates (see Details).\cr
\code{distance} \tab The distance map, a pixel image
(\code{\link[spatstat]{im.object}}).\cr
\code{mask} \tab A mask matrix showing the locations which are closer to
the tissue boundary than to the border or other regions.\cr
}
}
\description{
Given a cell seg table and an image containing masks for two tissue
classes, estimate the dependence of the density of cells of each
phenotype on the distance from the boundary between the two tissue classes.
}
\details{
The density estimate is computed by \code{\link[spatstat]{rhohat}}.
The signed distance from the boundary between the two tissue classes
is used as the covariate and density is estimated separately for each
phenotype.

The \code{rhohat} element of the returned list is a
\code{list} containing the results of the density
estimation for each phenotype. Each list value is a \code{rhohat} object,
see \code{\link[spatstat]{methods.rhohat}}. Density estimates are
in cells per square micron; multiply by 1,000,000 for cells per square
millimeter.
}
\section{Edge correction}{


The \code{\link[spatstat]{rhohat}} function does not have any built-in
edge correction. This may lead to incorrect density estimates
because it does not account for cells at the edge of the image which may
be near a tissue boundary which is not part of the image.

The \code{mask} parameter, if set to \code{TRUE}, restricts the density estimation
to cells which are closer to the tissue boundary in the image than they are
to the edge of the image or any other tissue class.

For both \code{mask==TRUE} and \code{mask==FALSE}, the results tend to be unreliable
near the distance extremes because the tissue area at that distance
will be relatively small so random variation in cell counts is magnified.
}

\section{Parallel computation}{


\code{density_at_distance} supports parallel computation using the
\code{\link[foreach]{foreach-package}}. When this is enabled, the densities
for each phenotype will be computed in parallel. To use this feature,
you must enable a parallel backend, for example using
\code{\link[doParallel]{registerDoParallel}}.
}

\examples{
# Compute density for the sample data
values <- density_at_distance(sample_cell_seg_path(),
  list("CD8+", "CD68+", "FoxP3+"),
  positive="Stroma", negative="Tumor")

# Combine all the densities into a single data_frame
# and filter out the extremes
library(tidyverse)
all_rho <- purrr::map_dfr(values$rhohat, ~., .id='phenotype') \%>\%
  as_data_frame \%>\% filter(X>=-50, X<=100)

# Plot the densities in a single plot
library(ggplot2)
ggplot(all_rho, aes(X, rho*1000000, color=phenotype)) +
  geom_line(size=2) +
  labs(x='Distance from tumor boundary (microns)',
       y='Estimated cell density (cells per sq mm)')

# Show the distance map with CD68+ cells superimposed
plot_diverging(values$distance, show_boundary=TRUE,
  title=paste('Distance from tumor for CD68+cells'),
  sub='Positive (blue) distances are away from tumor')
plot(values$points[values$points$marks=='CD68+', drop=TRUE],
  add=TRUE, use.marks=FALSE, cex=0.5, col=rgb(0,0,0,0.5))
}
\references{
A. Baddeley, E. Rubak and R.Turner.
Spatial Point Patterns: Methodology and Applications with R.
Chapman and Hall/CRC Press, 2015. Sections 6.6.3-6.6.4.
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{count_touching_cells}},
  \code{\link{count_within_batch}},
  \code{\link{count_within}}, \code{\link{density_bands}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{spatial_distribution_report}},
  \code{\link{subset_distance_matrix}}
}
