% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_bands.R
\name{density_bands}
\alias{density_bands}
\title{Estimate cell density in bands from a tissue boundary.}
\usage{
density_bands(cell_seg_path, phenotypes, positive, negative, width = 25,
  pixels_per_micron = getOption("phenoptr.pixels.per.micron"))
}
\arguments{
\item{cell_seg_path}{Path to a cell segmentation data file.}

\item{phenotypes}{Optional named list of phenotypes to process.
\code{names(phenotypes)} are the names of the resulting phenotypes.
The values are in any format accepted by \code{\link{select_rows}}.
If omitted, will use all phenotypes in the cell seg data.}

\item{positive}{Name of the tissue category used as positive distance,
e.g. "stroma".}

\item{negative}{Name of the tissue category used as negative distance,
e.g. "tumor".}

\item{width}{Width of the bands, in microns}

\item{pixels_per_micron}{Conversion factor to microns.}
}
\value{
Returns a \code{data_frame} with five columns:
\tabular{ll}{
\code{phenotype} \tab The supplied phenotypes.\cr
\code{midpoint} \tab The midpoint of the distance band.\cr
\code{count} \tab The number of cells of the phenotype found
within the band.\cr
\code{area} \tab The area of the band, in square microns.\cr
\code{density} \tab The density of cells of the phenotype in the band.\cr
}
}
\description{
Given a cell seg table and an image containing masks for two tissue
classes, estimate the density of cells of each
phenotype in bands from the boundary between the two
tissue classes.
}
\details{
The density estimate is computed by counting the number of cells
within each distance band, and dividing by the area of the band.
The returned value includes the cell count and area of each band,
making it straightforward to aggregate across images from a
single sample.
}
\examples{
# Compute density for the sample data
values <- density_bands(sample_cell_seg_path(),
  list("CD8+", "CD68+", "FoxP3+"),
  positive="Stroma", negative="Tumor")

# Plot the densities in a single plot
library(ggplot2)
ggplot(values, aes(midpoint, density*1000000, color=phenotype)) +
  geom_line(size=2) +
  labs(x='Distance from tumor boundary (microns)',
       y='Estimated cell density (cells per sq mm)')
}
\seealso{
Other distance functions: \code{\link{compute_all_nearest_distance}},
  \code{\link{count_touching_cells}},
  \code{\link{count_within_batch}},
  \code{\link{count_within}},
  \code{\link{density_at_distance}},
  \code{\link{distance_matrix}},
  \code{\link{find_nearest_distance}},
  \code{\link{spatial_distribution_report}},
  \code{\link{subset_distance_matrix}}
}
