% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Create a distance matrix from cell seg data.}
\usage{
distance_matrix(csd)
}
\arguments{
\item{csd}{A data frame with \code{Cell X Position}
and \code{Cell Y Position} columns, such as the result of
\code{\link{read_cell_seg_data}}.}
}
\value{
A square matrix with both dimensions equal to \code{nrow(d)}.
        The value at \code{[i, j]} will be the distance from the cell
        at row \code{i} of \code{csd} to the cell at row \code{j}.
        The returned matrix is symmetric.
}
\description{
Create a distance matrix from cell seg data.
}
\seealso{
Other distance functions: 
\code{\link{compute_all_nearest_distance}()},
\code{\link{count_touching_cells}()},
\code{\link{count_within_batch}()},
\code{\link{count_within_many}()},
\code{\link{count_within}()},
\code{\link{find_nearest_distance}()},
\code{\link{spatial_distribution_report}()},
\code{\link{subset_distance_matrix}()}
}
\concept{distance functions}
