% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{find_nearest_distance}
\alias{find_nearest_distance}
\title{For each phenotype in a single sample,
find the distance from
each cell to the nearest other cell in the phenotype.}
\usage{
find_nearest_distance(csd, phenotypes = NULL)
}
\arguments{
\item{csd}{A data frame with \code{Cell X Position},
\code{Cell Y Position} and \code{Phenotype} columns,
such as the result of calling
\code{\link{read_cell_seg_data}}.}

\item{phenotypes}{Optional list of phenotypes to include. If omitted,
\code{unique(csd$Phenotype)} will be used.}
}
\value{
A data_frame containing a 'Distance to <phenotype>' column
for each phenotype. Will contain NA values where there is no other cell
of the phenotype.
}
\description{
For each phenotype in a single sample,
find the distance from
each cell to the nearest other cell in the phenotype.
}
\examples{
# Compute distance columns and append them to the source data
d = sample_cell_seg_data
d = cbind(d, find_nearest_distance(d))
}

