% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_invert.R
\docType{data}
\name{geom_sf_invert}
\alias{geom_sf_invert}
\alias{StatSfInvert}
\alias{stat_sf_invert}
\alias{scale_sf_invert}
\alias{negate_y}
\title{Negate the y-axis of a simple features object}
\usage{
geom_sf_invert(stat = StatSfInvert, ...)

scale_sf_invert(...)

negate_y(sf_obj)
}
\arguments{
\item{stat}{The statistical transformation to use on the data for this layer,
as a string or Stat object.}

\item{...}{Passed on to \code{ggplot2::geom_sf()} or \code{ggplot2::scale_y_continuous()}}

\item{sf_obj}{A simple features object}
}
\value{
\code{negate_y()} returns \code{sf_obj} with the y-axis negated.
}
\description{
Use \code{geom_sf_invert()} with \code{scale_sf_invert()} to match the orientation
of Polaris images.
}
\keyword{datasets}
