% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_components.R
\name{get_field_info}
\alias{get_field_info}
\title{Read information about a field from a component image file.}
\usage{
get_field_info(path)
}
\arguments{
\item{path}{Path to the \code{component_data.tif} file.}
}
\value{
A named list with values
\describe{
\item{\code{location}}{The physical (x, y) location of the top-left
corner of the field on the source slide, in microns.}
\item{\code{field_size}}{The physical (width, height)
of the field, in microns.}
\item{\code{image_size}}{The (width, height) of the field image,
in pixels.}
\item{\code{microns_per_pixel}}{The size of each pixel, in microns.}
}
}
\description{
Find the location, size and magnification of an inForm field by inspecting a
\code{component_data.tif} file.
}
\details{
The field location is determined from the coordinates in the file
name. The field size and magnification are read from TIFF tags
in the file.

Limited to 1x1 fields due to
limitations in the \code{tiff} package. Install the PerkinElmer
\href{https://github.com/PerkinElmer/tiff}{fork} of the tiff package to
remove this limitation.
}
\examples{
if (require('phenoptrExamples')) {
  path <- system.file('extdata', 'samples',
    'Set4_1-6plex_[16142,55840]_component_data.tif',
    package='phenoptrExamples')
  get_field_info(path)
}
}
\seealso{
Other file readers: \code{\link{list_cell_seg_files}},
  \code{\link{read_cell_seg_data}},
  \code{\link{read_components}}, \code{\link{read_maps}}
}
\concept{file readers}
