% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{parse_rect}
\alias{parse_rect}
\title{Parse a single rectangle annotation}
\usage{
parse_rect(field, ns = xml2::xml_ns(field))
}
\arguments{
\item{field}{Parsed XML for a single rectangle annotation}

\item{ns}{Optional namespace; providing this makes parsing many rects
much faster.}
}
\value{
An \code{sf::st_sf} with coordinates of the rectangle
and columns \code{center_x} and \code{center_y} giving the \emph{rounded}
center point that inForm uses to identify the field.
}
\description{
Parse a single rectangle annotation
}
\keyword{internal}
