% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_at_distance.R
\name{plot_diverging}
\alias{plot_diverging}
\title{Plot a distance map using a diverging color scale with white at 0}
\usage{
plot_diverging(im, title, show_boundary = FALSE, ...)
}
\arguments{
\item{im}{A pixel image of class \code{\link[spatstat.geom]{im}}.}

\item{title}{Title for the plot.}

\item{show_boundary}{Should the boundary be highlighted?}

\item{...}{Additional arguments passed to \code{\link[spatstat.geom]{plot.im}}}
}
\description{
Plot a distance map using a diverging color scale with white at 0
}
