% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cell_seg_data.R
\name{read_cell_seg_data}
\alias{read_cell_seg_data}
\title{Read and clean an inForm data file.}
\usage{
read_cell_seg_data(path = NA,
  pixels_per_micron = getOption("informr.pixels.per.micron"),
  remove_units = TRUE)
}
\arguments{
\item{path}{Path to the file to read, or NA to use a file chooser.}

\item{pixels_per_micron}{Conversion factor to microns
(default 2 pixels/micron),
set to NA to skip conversion.}

\item{remove_units}{If TRUE (default),
remove the unit name from expression columns.}
}
\value{
A \code{\link[tibble]{data_frame}}
        containing the cleaned-up data set.
}
\description{
\code{read_cell_seg_data} makes it easier to use data from PerkinElmer's
inForm program. It reads data files written by inForm 2.0 and later and does
useful cleanup on the result. Data files written by inForm 2.0 can be read
easily using \code{\link[utils]{read.delim}} or
\code{\link[readr]{read_tsv}}. However there is still some useful cleanup to
be done.
}
\details{
\code{read_cell_seg_data} reads both single-image tables and merged tables
and does useful cleanup on the data:
\itemize{
\item Removes columns that are all NA
      (these are typically unused summary columns)
\item Converts percent columns to numeric fractions
\item Converts pixel distances to microns. The conversion factor may be
specfied as a parameter or by setting
\code{options(informr.pixels.per.micron)}.
\item Optionally removes units from expression names
\item If the file contains multiple sample names,
      a \code{tag} column is created
      containing a minimal, unique tag for each sample.
      This is useful when a
      short name is needed, for example in chart legends.
}
}
\examples{
path = system.file("extdata", "TMA",
                   "Core[1,5,6,1]_[21302,15107]_cell_seg_data.txt",
                   package = "informr")
d = read_cell_seg_data(path)

\dontrun{
# Read all cell seg files in a directory and return a single data_frame
paths = list_cell_seg_files(path)
d = purrr::map_df(paths, read_cell_seg_data)
}
}
\seealso{
Other file readers: \code{\link{list_cell_seg_files}},
  \code{\link{read_components}}, \code{\link{read_maps}}
}
