% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maps.R
\name{read_maps}
\alias{read_maps}
\title{Read an inForm segmentation map file.}
\usage{
read_maps(map_path)
}
\arguments{
\item{map_path}{Path to the map file or a cell seg data file in the same
directory.}
}
\value{
A named list of images, one for each map in the source file.
}
\description{
Reads an inForm `binary_seg_maps.tif`` file and returns a named
list of images. The names reflect the content of the individual images.
Possible names are Nucleus, Cytoplasm, Membrane, Object,
TissueClassMap, and ProcessRegionImage; not every image file will include all
names.
}
\details{
Images are oriented to match the coordinates in a cell seg data file,
i.e. (0, 0) at the top left and the row corresponding to Y and
column corresponding to X.
}
\examples{
\dontrun{
# Not included at this time
path = system.file("extdata", "TMA",
                   "Core[1,5,6,1]_[21302,15107]_binary_seg_maps.tif",
                   package = "phenoptr")
d = read_maps(path)
names(d)
}
}
\seealso{
Other file readers: \code{\link{list_cell_seg_files}},
  \code{\link{read_cell_seg_data}},
  \code{\link{read_components}}
}
