% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{read_phenochart_polygons}
\alias{read_phenochart_polygons}
\title{Read polygons (ROIs) and tags from a Phenochart annotation.}
\usage{
read_phenochart_polygons(xml_path)
}
\arguments{
\item{xml_path}{Path to an annotations file.}
}
\value{
An \code{sf::st_sf} object with columns \code{tags}, \code{geometry} and \code{rects}.
Multiple tags for a single ROI are separated by spaces in a single string.
}
\description{
Reads the polygons and included rectangles for all ROI annotations
in a single file.
}
