% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{read_phenochart_polygons}
\alias{read_phenochart_polygons}
\title{Read polygons and tags from a Phenochart annotation.}
\usage{
read_phenochart_polygons(xml_path)
}
\arguments{
\item{xml_path}{Path to an annotations file.}
}
\value{
An \code{sf::sf} object with columns \code{tags} and \code{geometry}. Multiple
tags are separated by spaces in a single string.
}
\description{
Reads the polygons for all ROI annotations in a single file.
}
