% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readTIFFDirectory}
\alias{readTIFFDirectory}
\title{Read directory info from a TIFF file}
\usage{
readTIFFDirectory(path, all = FALSE)
}
\arguments{
\item{path}{Path to a TIFF file}

\item{all}{Return info on all images or just the first one?}
}
\value{
A list (if \code{all==FALSE}) or list of lists (if \code{all==TRUE}) containing
directory info for the images in the file
}
\description{
Read directory info from a TIFF file
}
