% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cell_seg_data}
\alias{sample_cell_seg_data}
\title{Sample cell segmentation data}
\format{A data frame with 5726 rows and 200 variables}
\usage{
sample_cell_seg_data
}
\description{
A dataset containing inForm cell segmentation data for a single TMA core.
This is the data from
\code{system.file("extdata",
                  "TMA/Core[1,5,6,1]_[21302,15107]_cell_seg_data.txt",
                 package = "informr")}
}
\details{
To see all the included files for this core, use
\code{list.files(system.file("extdata", "TMA", package = "informr"))}
}
\keyword{datasets}
