% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_cell_seg_data}
\alias{sample_cell_seg_data}
\alias{sample_cell_seg_path}
\alias{sample_cell_seg_folder}
\title{Sample cell segmentation data.}
\format{
A data frame with 6072 rows and 199 variables
}
\usage{
sample_cell_seg_data

sample_cell_seg_path()

sample_cell_seg_folder()
}
\description{
A dataset containing inForm cell segmentation data for a single field.
}
\details{
\code{sample_cell_seg_data} contains the \code{data_table} returned from\if{html}{\out{<div class="r">}}\preformatted{read_cell_seg_data(sample_cell_seg_path())
}\if{html}{\out{</div>}}

This table shows the stains used, the epitopes they were bound to,
and the colors used to show the stains in the composite and phenotype views.\preformatted{  Stain    Epitope  Composite  Phenotype
  -------- -------  ---------  ---------
  DAPI     Nucleus  Blue       N/A
  Opal520  PDL1     Red        N/A
  Opal540  CD8      Yellow     Yellow
  Opal570  FoxP3    Orange     Orange
  Opal620  CD68     Magenta    Magenta
  Opal650  PD1      Green      N/A
  Opal690  CK       Cyan       Cyan
}

The sample cell data includes phenotypes \verb{CD8+} (cytotoxic T cell),
\verb{CD68+} (macrophage),
\verb{FoxP3+} (regulatory T cell), \verb{CK+} (tumor), and \code{other}.

\link{sample_cell_seg_path} and \link{sample_cell_seg_folder} are
convenience functions which return the path to the sample cell seg file
included in this package, or the folder containing
sample cell seg data, respectively.

To see all the files included for this field, use
\code{list.files(sample_cell_seg_folder())}.

For more sample data from this and related samples, see the
\href{https://akoyabio.github.io/phenoptrExamples}{phenoptrExamples} package.
}
\examples{
dim(sample_cell_seg_data)
table(sample_cell_seg_data$Phenotype)
}
\keyword{datasets}
