% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rows.R
\name{select_rows}
\alias{select_rows}
\title{Flexibly select rows of a data frame.}
\usage{
select_rows(csd, sel)
}
\arguments{
\item{csd}{A data frame}

\item{sel}{May be a character vector, a one-sided formula, a list
containing such or \code{NA}. A character vector is interpreted as
the name(s) of one or
more phenotypes or hashtags and selects any matching phenotype.
A formula is interpreted as an expression on the columns of \code{csd}.
Multiple list items are joined with AND. \code{NA} is interpreted
as "select all". It is convenient for lists of selection criteria.}
}
\value{
A logical vector of length \code{nrow(csd)} which selects rows
according to \code{sel}.
}
\description{
Select rows of a data frame based on phenotypes or other
expressions.
}
\details{
\code{select_rows} implements a flexible mechanism for selecting cells (rows)
from a cell segmentation table. Cells may be selected by single or
multiple phenotype, by expression level, by hashtag,
or combinations of all three.

See the tutorial
\href{https://akoyabio.github.io/phenoptr/articles/selecting_cells.html}{Selecting cells within a cell segmentation table}
for extensive documentation and examples.
}
\examples{
csd <- sample_cell_seg_data

# Select tumor cells with PDL1 expression > 3
selector <- list('CK+', ~`Entire Cell PDL1 (Opal 520) Mean`>3)
pdl1_pos_tumor <- csd[select_rows(csd, selector),]
range(pdl1_pos_tumor$`Entire Cell PDL1 (Opal 520) Mean`)

# Select all T-cells. Note: Use c() to combine phenotypes, not list()
selector <- c('CD8+', 'FoxP3+')
tcells <- csd[select_rows(csd, selector),]
table(tcells$Phenotype)
}
\seealso{
\link{parse_phenotypes} for a convenient way to create selectors
for most common phenotypes.
}
